Move REST API to its own class
