+++ RestAPI/MolTwisterRestAPI.h
@@ -1,0 +1,13 @@
+#pragma once
+
+class CMolTwisterRestAPI
+{
+public:
+    CMolTwisterRestAPI() = default;
+
+public:
+    void run() const;
+
+private:
+    static void* threadRun(void*);
+};
